+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 1.1.3-preview-4
 ### Improvement
 - Session log now shows the version of tasmota-webserial-esptool package along with a timestamp
+- Failed detection of flashid is now non-fatal
 
 ### Internal
-- Failed detection of flashid is now non-fatal
+- Serial Monitor start now use native tasmota-webserial-esptool hardreset(false) to enter user firmware
 - Bumps tasmota-webserial-esptool from 7.2.2 to 7.2.3
 - Bumps vuetify from 3.11.4 to 3.11.5
 - Bumps vue-i18n from 9.14.5 to 11.2.7.